Rename LabelText to Label and adjust examples
