+++ src/core/Form/Label/Label.md
@@ -1,0 +1,11 @@
+Label is an accessible label component to use with custom input elements. Applicable form components in the library already have Label as an integral part.
+
+```js
+import { Label } from 'suomifi-ui-components';
+import React from 'react';
+<>
+  <Label>This is a label for an adjacent input</Label>
+  <Label optionalText="optional">Label for an optional field</Label>
+  <Label labelMode="hidden">A visually hidden label</Label>
+</>;
+```
