+++ src/core/Form/Label/Label.test.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { axeTest } from '../../../utils/test';
 
-import { LabelText } from './LabelText';
+import { Label } from './Label';
 
 describe('props', () => {
   describe('children', () => {
     it('shows the given text', () => {
-      const { container } = render(<LabelText>Test text</LabelText>);
+      const { container } = render(<Label>Test text</Label>);
       expect(container.firstChild).toHaveTextContent('Test text');
     });
   });
@@ -15,7 +15,7 @@ describe('props', () => {
   describe('optionalText', () => {
     it('has the optional text element', () => {
       const { getByText } = render(
-        <LabelText optionalText="optional">Test text</LabelText>,
+        <Label optionalText="optional">Test text</Label>,
       );
       const optionalText = getByText('(optional)');
       expect(optionalText).toHaveClass('fi-label-text_optionalText');
@@ -24,9 +24,7 @@ describe('props', () => {
 
   describe('id', () => {
     it('has the given id', () => {
-      const { container } = render(
-        <LabelText id="test-id">Test text</LabelText>,
-      );
+      const { container } = render(<Label id="test-id">Test text</Label>);
       expect(container.firstChild).toHaveAttribute('id', 'test-id');
     });
   });
@@ -34,7 +32,7 @@ describe('props', () => {
   describe('className', () => {
     it('has the given custom classname', () => {
       const { container } = render(
-        <LabelText className="custom-style">Test text</LabelText>,
+        <Label className="custom-style">Test text</Label>,
       );
       expect(container.firstChild).toHaveClass('custom-style');
     });
@@ -42,15 +40,13 @@ describe('props', () => {
 
   describe('labelMode', () => {
     it('should be visible by default', () => {
-      const { getByText } = render(<LabelText>Test text</LabelText>);
+      const { getByText } = render(<Label>Test text</Label>);
       const label = getByText('Test text');
       expect(label).toHaveClass('fi-label-text_label-span');
     });
 
     it('should be hidden', () => {
-      const { getByText } = render(
-        <LabelText labelMode="hidden">Test text</LabelText>,
-      );
+      const { getByText } = render(<Label labelMode="hidden">Test text</Label>);
       const label = getByText('Test text');
       expect(label).toHaveClass('fi-visually-hidden');
     });
@@ -58,14 +54,12 @@ describe('props', () => {
 
   describe('asProp', () => {
     it('has default of label as wrapping element', () => {
-      const { container } = render(<LabelText>Test text</LabelText>);
+      const { container } = render(<Label>Test text</Label>);
       expect(container.firstChild).toBeInstanceOf(HTMLLabelElement);
     });
 
     it('has the given wrapper element', () => {
-      const { container } = render(
-        <LabelText asProp="div">Test text</LabelText>,
-      );
+      const { container } = render(<Label asProp="div">Test text</Label>);
       expect(container.firstChild).toBeInstanceOf(HTMLDivElement);
     });
   });
@@ -73,9 +67,7 @@ describe('props', () => {
   describe('labelSpanProps', () => {
     it('has the given props', () => {
       const { getByText } = render(
-        <LabelText labelSpanProps={{ style: { fontSize: 12 } }}>
-          Test text
-        </LabelText>,
+        <Label labelSpanProps={{ style: { fontSize: 12 } }}>Test text</Label>,
       );
       const textSpan = getByText('Test text');
       expect(textSpan).toHaveAttribute('style', 'font-size: 12px;');
@@ -85,5 +77,5 @@ describe('props', () => {
 
 test(
   'should not have basic accessibility issues',
-  axeTest(<LabelText>Test text</LabelText>),
+  axeTest(<Label>Test text</Label>),
 );
