+++ src/core/Form/index.ts
@@ -22,6 +22,6 @@ export {
   SingleSelectData,
   SingleSelectStatus,
 } from './Select';
-export { LabelText, LabelTextProps } from './LabelText/LabelText';
+export { Label, LabelProps } from './Label/Label';
 export { StatusText, StatusTextProps } from './StatusText/StatusText';
 export { HintText, HintTextProps } from './HintText/HintText';
